requirements.txt
python-dotenv
requests
pandas
numpy
urllib3
pycurl
schedule